Add dropdown menu to show services
I use a 'mouseover' event listener to show a dropdown menu. I decided
to keep the dropdown menu and the element 'servicios' that triggers the
event separated in order to maintain the existing document flow. To
hide the dropdown menu, I added a 'mouseout' event to the dropdown
itself. This cause an undesired behavior, if accidentally the user
hovered over 'servicios' then the dropdown would stay visible, if the
mouse never hovered over it.

To fix this, I used a wrapper that slightly overlaps with 'servicios'.
Then I added both the 'mouseover' and 'mouseout' events to both,
'servicios' and the wrapper. This allows the dropdown to stay visible
when the mouse crosses from 'servicios' and prevents it to remain
visible if the user accidentally hovers over the 'servicios' tab.